Move StateTurnContextExtensions into own file, more comments
