+++ samples-final/8.AspNetCore-LUIS-Bot/MyBotAccessors.cs
@@ -8,7 +8,11 @@ namespace AspNetCore_LUIS_Bot
 {
     /// <summary>
     /// Creates the State Accessors used by the LuisBot. In general usage, this class
-    /// is created as a Singleton and passed into the LuisBot constructor
+    /// is created as a Singleton and passed into the IBot-derived LuisBot constructor.
+    ///  - See MyLuisBot.cs constructor for how that is injected 
+    ///  - See the Startup.cs file for more details on creating the Singleton that gets
+    ///    injected into the constructor.
+    /// 
     /// </summary>
     public class MyBotAccessors
     {
